If statement completed
The If statement has been completed.

All three variables (previousCharacter, currentCharacter, and previousPlusCurrent) have been instantiated as "" and are set to actual values at the beginning of the for loop. I did this to prevent the program from processing the same set of variables (k=0) twice. Alternatively, I could have left the variables in their original place and just changed the loop to start at 1 rather than 0.