decompression
basically, I did it without needing to input the dictionary. im ngl, it was a lot of trial and error and my head hurts. but basically u can reason out the dictionary from the input file? I tried to make the comments coherent but it's kinda late so idk if they are haha.